% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cran_packages}
\alias{get_cran_packages}
\title{Return all CRAN packages installed from the user}
\usage{
get_cran_packages()
}
\value{
A tibble with name, version and source
}
\description{
Return all CRAN packages installed from the user
}
