% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_github_packages}
\alias{get_github_packages}
\title{Return all Github packages installed from the user.}
\usage{
get_github_packages()
}
\value{
A tibble with name, version and source
}
\description{
Return all Github packages installed from the user.
}
