% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_github_source}
\alias{get_github_source}
\title{Transform the source column from get_github_packages for using the string
directly on \code{devtools::install_github()}}
\usage{
get_github_source(github_packages)
}
\value{
a character vector with all github sources
}
\description{
Transform the source column from get_github_packages for using the string
directly on \code{devtools::install_github()}
}
