% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_pkgs}
\alias{restore_pkgs}
\title{Install all packages contained in the \code{rds} file created with
\code{save_installed_packages()}}
\usage{
restore_pkgs(file)
}
\arguments{
\item{file}{the \code{rds} file with all packages}
}
\value{
None
}
\description{
Install all packages contained in the \code{rds} file created with
\code{save_installed_packages()}
}
