% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_installed_packages}
\alias{save_installed_packages}
\title{Save an \code{.rds} file with all packages in the specified path}
\usage{
save_installed_packages(file = "pkgs_list.rds")
}
\arguments{
\item{file}{path to create the \code{rds} file}
}
\value{
None
}
\description{
Save an \code{.rds} file with all packages in the specified path
}
